backend/requirements.txt
dj-database-url==0.5.0
djangorestframework==3.12.1
Django==3.0.8
whitenoise==5.1.0
psycopg2==2.8.6
pre-commit==2.11.1